Problem set 3
Context: Consider the dataset 'Impact of COVID-19 on worldwide aviation', which concerns the impact of global pandemic on the airline industry.

Task: Create ONE Matplotlib figure that illustrates the effect of the global pandemic on the airline industry. There are no constraints about the dimensions to include in the figure ― in other words, students are free to emphasize the aspects they find more interesting. However, the plot(s) included in the Matplotlib figure(s) must include the temporal dimension.